# packages/shared/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: workspace-chat API
  version: 0.0.0
  description: >-
    エラーの応答の本体はすべて ErrorResponse である。仕様に無いパス（前置き /api の外を含む）は
    404（not_found）を返し、パスごとには載せない。
servers:
  # 全ルートの前置きは /api（#77 の決定。apps/api/src/app-setup.ts の createApp が付ける）。
  - url: /api
# 既定ですべての操作にアクセストークンを求める。認証を要さない操作だけが `security: []` で外す
# （api 側は auth/access-token.guard.ts が同じ既定を持つ。一致は auth/access-token.test.ts が仕様を読んで確かめる）。
security:
  - bearerAuth: []
paths:
...
  /workspaces/{id}/channels/{channelId}/messages:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/ChannelId'
    get:
      operationId: listMessages
      summary: チャンネルのメッセージの一覧（F-11・F-12）
      description: >-
        新しい順。before（メッセージの id）を渡すと、それより古いものを返す（カーソルページネーション。OFFSET を使わない。機能一覧 4.1）。
        スレッドの返信は含めない（返信の一覧は replies。機能一覧 6）。
        before はそのチャンネルのメッセージでなくてもよく、存在を確かめない（id の順で比べる境目にすぎない）。
        続きがあれば nextBefore に次の before を、無ければ null を返す。
        参加者だけが取得できる。コードは2段階で決まる: 所属していなければ種別によらず 404。所属していて参加していなければ、
        パブリックは 403 not_a_channel_member・プライベートは 404。オーナーでも参加していなければ同じ（オーナーの例外はメッセージに及ばない。機能一覧 3.1・4.1）。
        別のワークスペースのチャンネル・無いチャンネルは 404。アーカイブ済みでも参加者は読める（機能一覧 3.2）。
        退会した投稿者のメッセージは残し、author を null にする（機能一覧 1.5）
      parameters:
        - name: before
          in: query
          required: false
          description: これより古いメッセージを返す（メッセージの id）。形が uuid でなければ 400
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          description: 返す件数の上限。1〜100、既定は 50（実装時に決めた値。機能一覧 4.1）
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: メッセージ（新しい順）と、続きを取るための before
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: パブリックチャンネルに参加していない（not_a_channel_member）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: postMessage
      summary: チャンネルへのメッセージの投稿（F-11）
      description: >-
        参加者だけが投稿できる。コードは一覧と同じ2段階（所属していなければ 404。所属していて参加していなければ、パブリックは 403 not_a_channel_member・
        プライベートは 404。オーナーでも参加していなければ同じ）。アーカイブ済みのチャンネルには投稿できない（409 channel_archived。機能一覧 3.2）。
        本文は 1〜4000 文字（コードポイント）で、空白だけは不可（機能一覧 4.1）。
        作ったメッセージを message:new としてチャンネルの部屋へ配る。本文のメンションの対象がいれば、その利用者の部屋も加えて1回で配る（機能一覧 5.2・9.1）。
        attachmentIds で確定した添付を付けられる（自分が上げ、このチャンネルで確定に成功し、まだどの投稿にも付いていないものだけ。そうでなければ 422 attachment_unavailable で投稿しない。機能一覧 11.1）。
        投稿・返信・編集・削除を合わせて、利用者単位で1分に60回まで（4つのルートで1つの枠。機能一覧 4.1・4.2・6）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageRequest'
      responses:
        '201':
          description: 作ったメッセージ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: パブリックチャンネルに参加していない（not_a_channel_member）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '409':
          description: アーカイブ済みのチャンネル（channel_archived）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          description: 付けられない添付がある（attachment_unavailable）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /workspaces/{id}/dms:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    get:
      operationId: listDms
      summary: 自分が当事者の DM の一覧（F-19・F-23）
      description: >-
        そのワークスペースで、要求した利用者が当事者の DM だけを返す（機能一覧 8）。所属していなければ 404。
        新しいメッセージのある順（メッセージの無い DM は始めた順の位置に並ぶ）。
        相手がキック・退出・退会した DM も返す（過去のメッセージは残る。機能一覧 1.5・2.2）——相手が退会していれば counterpart は null、
        いまメンバーでなければ writable は false（投稿できない）
      responses:
        '200':
          description: DM（新しいメッセージのある順）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dm'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: startDm
      summary: DM を始める・開く（F-19）
      description: >-
        相手（User.id）との DM を返す。まだ無ければ作る。同じ相手との DM は1つに集約され、既にあればそれを返す（機能一覧 8）。
        相手は同じワークスペースの、退会していないメンバーに限る（そうでなければ 422 dm_counterpart_not_found）。
        自分自身とは始められない（422 dm_with_self）。所属していなければ 404
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartDmRequest'
      responses:
        '200':
          description: 相手との DM（作ったもの、または既にあったもの）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dm'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          description: 相手がこのワークスペースの退会していないメンバーでない（dm_counterpart_not_found）か、自分自身（dm_with_self）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workspaces/{id}/dms/{dmId}/messages:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/DmId'
    get:
      operationId: listDmMessages
      summary: DM のメッセージの一覧（F-19）
      description: >-
        新しい順。before・limit・nextBefore の扱いはチャンネルの一覧と同じ（機能一覧 4.1）。
        当事者だけが取得できる。所属していなければ 404、当事者でない DM・無い DM・別のワークスペースの DM も 404（存在の有無を区別しない。機能一覧 8）。
        相手がいまメンバーでなくても、当事者は過去のメッセージを読める（機能一覧 1.5・2.2）。
        退会した書き手のメッセージは author を null にする（機能一覧 1.5）
      parameters:
        - name: before
          in: query
          required: false
          description: これより古いメッセージを返す（メッセージの id）。形が uuid でなければ 400
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          description: 返す件数の上限。1〜100、既定は 50（チャンネルの一覧と同じ。機能一覧 4.1）
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: メッセージ（新しい順）と、続きを取るための before
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DmMessagePage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: postDmMessage
      summary: DM へのメッセージの投稿（F-19）
      description: >-
        当事者だけが投稿できる（所属していない・当事者でない・無い DM は 404）。
        相手がいまそのワークスペースの退会していないメンバーでなければ 409 dm_counterpart_unavailable（機能一覧 8「相手は同一ワークスペースのメンバーに限る」）。
        本文の形はチャンネルの投稿と同じ（1〜4000 文字・空白だけは不可）。
        作ったメッセージを message:new として、当事者2人のうち、いまそのワークスペースの退会していないメンバーである利用者の部屋へ1回で配る（機能一覧 5.2 の DM の箇条）。
        投稿・返信・編集・削除を合わせて、利用者単位で1分に60回まで（チャンネルのメッセージの書き込みと同じ枠。機能一覧 4.1）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostMessageRequest'
      responses:
        '201':
          description: 作ったメッセージ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DmMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '409':
          description: 相手がいまこのワークスペースの退会していないメンバーでない（dm_counterpart_unavailable）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workspaces/{id}/dms/{dmId}/messages/{messageId}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/DmId'
      - $ref: '#/components/parameters/MessageId'
    patch:
      operationId: editDmMessage
      summary: DM のメッセージの編集（F-19・F-13）
      description: >-
        自分のメッセージだけを編集できる。判定の順: 所属していなければ 404 → 当事者でない・無い DM は 404 →
        その DM に無い・削除済みのメッセージは 404 → 作者でなければ 403 not_message_author。
        相手がいまメンバーでなくても、自分のメッセージは編集できる。本文の形は投稿と同じ。
        編集したメッセージを message:updated として、投稿と同じ宛先へ配る（機能一覧 5.2 の DM の箇条）。
        投稿・返信・編集・削除を合わせて、利用者単位で1分に60回まで（機能一覧 4.1・4.2）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostMessageRequest'
      responses:
        '200':
          description: 編集したメッセージ（editedAt を持つ）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DmMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: 自分のメッセージではない（not_message_author）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: deleteDmMessage
      summary: DM のメッセージの削除（F-19・F-13）
      description: >-
        論理削除（機能一覧 4.2）。自分のメッセージだけを削除できる。判定の順は編集と同じ。本文は DB に残るが、
        以後の応答と配信では返さない（body: null・deleted: true）。message:deleted（本文を載せない）を投稿と同じ宛先へ配る。
        投稿・返信・編集・削除を合わせて、利用者単位で1分に60回まで（機能一覧 4.1・4.2）
      responses:
        '204':
          description: 削除した
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: 自分のメッセージではない（not_message_author）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workspaces/{id}/dms/{dmId}/read:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/DmId'
    put:
      operationId: updateDmRead
      summary: DM の既読位置の更新（F-23）
      description: >-
        読んだ位置（その DM のメッセージの id）を渡して既読位置を進める（「利用者 × DM」の既読位置。機能一覧 10.1）。
        既読位置は戻さない——渡した id が既に読んだ位置より古ければ、何も変えずに 204 を返す。
        所属していない・当事者でない・無い DM は 404。その DM に無いメッセージ・削除済みのメッセージの id も 404。
        進めた後の未読数を unread:updated として本人の部屋へだけ配る（機能一覧 5.2）。
        利用者ごとに1分 120 回まで（チャンネル・スレッドの既読の更新とは別の枠。機能一覧 10.1）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDmReadRequest'
      responses:
        '204':
          description: 既読位置を進めた（進まなかった場合も 204）
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      description: ログイン（/auth/login）とリフレッシュ（/auth/refresh）が返すアクセストークン（JWT・HS256）
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ChannelId:
      name: channelId
      in: path
      required: true
      description: チャンネルの id。形が uuid でなければ 400
      schema:
        type: string
        format: uuid
    MessageId:
      name: messageId
      in: path
      required: true
      description: メッセージの id。形が uuid でなければ 400
      schema:
        type: string
        format: uuid
    DmId:
      name: dmId
      in: path
      required: true
      description: DM の id。形が uuid でなければ 400
      schema:
        type: string
        format: uuid
...
    WorkspaceId:
      name: id
      in: path
      required: true
      description: ワークスペースの id。形が uuid でなければ 400
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: >-
        Authorization ヘッダーが無い、または Bearer 方式でない（authentication_required）か、
        Bearer 方式で資格情報が無い・壊れている・期限切れ・利用者が退会済み（invalid_token。どれに当たったかは区別しない。機能一覧 1.4）。
        アクセストークンはリフレッシュ（/auth/refresh）で取り直す。仕様の形の検証（400）はトークンの確認より先に行う（仕様で書けない検証は後に行う）
      headers:
        WWW-Authenticate:
          description: >-
            Authorization ヘッダーが無い・Bearer 方式でないときは `Bearer`、Bearer 方式で資格情報が無い・使えないときは `Bearer error="invalid_token"`（RFC 6750 3.1）
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BadRequest:
      description: >-
        入力が仕様に合わない（validation_failed）か、本体を JSON として読めない（invalid_body）。
        どちらも送られた値を応答に載せない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    MethodNotAllowed:
      description: 仕様に無いメソッド（method_not_allowed）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PayloadTooLarge:
      description: 本体が大きすぎる（payload_too_large。上限は Nest の JSON の読み取りの既定の 100kb）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnsupportedMediaType:
      description: 仕様に無いメディア型の本体（unsupported_media_type）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: レート制限を超えた（too_many_requests。上限はパスの説明に書く）
      headers:
        Retry-After:
          description: 再び受け付けるまでの秒数
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: >-
        見つからない（not_found）。所属していないワークスペースは、存在の有無を区別せずこれになる（機能一覧 2.1・1.4）。
        本体は状態コードの本体であり、権限の有無を述べない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: 想定外の失敗（internal_error）。例外のメッセージは載せない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          description: >-
            エラーの種類。api が返す値はこの列挙だけであり、api と web は生成した型で同じ列挙を使う
            （綴りを誤ると型検査で落ちる）
          type: string
          enum:
            - validation_failed
            - invalid_body
            - not_found
            - method_not_allowed
            - payload_too_large
            - unsupported_media_type
            - too_many_requests
            - user_id_taken
            - registration_disabled
            - invalid_credentials
            - authentication_required
            - invalid_token
            - csrf_rejected
            - owner_only
            - invitee_not_found
            - already_invited
            - already_member
            - owner_cannot_leave
            - password_mismatch
            - owner_cannot_delete_account
            - pin_limit_reached
            - channel_name_taken
            - not_a_channel_member
            - already_channel_member
            - channel_archived
            - channel_not_private
            - channel_not_archived
            - not_message_author
            - dm_with_self
            - dm_counterpart_not_found
            - dm_counterpart_unavailable
            - reaction_limit_reached
            - unsupported_file_type
            - file_too_large
            - upload_not_received
            - upload_in_progress
            - attachment_unavailable
            - request_rejected
            - internal_error
        message:
          type: string
        errors:
          description: 入力の検証で落ちた箇所。送られた値は含めない
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - path
              - message
            properties:
              path:
                type: string
              message:
                type: string
...
    UserSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - userId
        - displayName
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        displayName:
          type: string
...
    CreateMessageRequest:
      description: >-
        チャンネルへの投稿の本体（機能一覧 4.1）。本文は PostMessageRequest と同じ形で、確定した添付を最大 10 件付けられる
      type: object
      additionalProperties: false
      required:
        - body
      properties:
        body:
          description: 1〜4000 文字（文字数はコードポイントで数える）。空白だけは不可（機能一覧 4.1）
          type: string
          minLength: 1
          maxLength: 4000
          pattern: '\S'
        attachmentIds:
          description: >-
            付ける添付の識別子（確定の応答の id）。自分が上げ、このチャンネルで確定に成功し、まだどの投稿にも付いていないものだけを付けられる
            （1つでも当たらなければ 422 attachment_unavailable で、投稿しない）。並びは上げた順になる
          type: array
          maxItems: 10
          uniqueItems: true
          items:
            type: string
            format: uuid
    Attachment:
      description: 添付ファイル（機能一覧 11.1・11.2）
      type: object
      additionalProperties: false
      required:
        - id
        - fileName
        - contentType
        - kind
        - size
        - url
      properties:
        id:
          description: アップロードの識別子
          type: string
          format: uuid
        fileName:
          description: 利用者が付けた元のファイル名（表示用。保存名とは違う）
          type: string
        contentType:
          description: 配信の Content-Type（検証した形式からサーバーが決めた値。テキスト系は text/plain; charset=utf-8）
          type: string
        kind:
          description: 種別（画像は表示し、動画は #t=30 の位置を表示し、それ以外はリンクにする。機能一覧 11.1）
          type: string
          enum:
            - image
            - video
            - document
            - archive
        size:
          description: 大きさ（バイト）
          type: integer
          minimum: 1
        url:
          description: >-
            配信 URL のパス（/files/workspace/{workspaceId}/channel/{channelId}/{id}/{保存名}）。CloudFront の署名付き Cookie で取得する（機能一覧 11.2）
          type: string
    PostMessageRequest:
      type: object
      additionalProperties: false
      required:
        - body
      properties:
        body:
          description: 1〜4000 文字（文字数はコードポイントで数える）。空白だけは不可（機能一覧 4.1）
          type: string
          minLength: 1
          maxLength: 4000
          pattern: '\S'
    Message:
      description: チャンネルのメッセージ（機能一覧 4.1・4.2）
      type: object
      additionalProperties: false
      required:
        - id
        - channelId
        - author
        - body
        - createdAt
        - editedAt
        - deleted
        - parentId
        - replyCount
        - replyParticipants
        - mentions
        - reactions
        - attachments
      properties:
        id:
          description: UUIDv7。一覧のカーソル（before）にそのまま使う
          type: string
          format: uuid
        channelId:
          type: string
          format: uuid
        parentId:
          description: スレッドの親の id。チャンネルの本体のメッセージなら null（機能一覧 6）
          type:
            - string
            - 'null'
          format: uuid
        replyCount:
          description: 削除されていない返信の件数。返信なら常に 0（返信はスレッドを持たない。機能一覧 6）
          type: integer
          minimum: 0
        replyParticipants:
          description: >-
            返信した参加者。削除されていない返信を書いた、退会していない利用者を、最後に返信した順に最大3人（機能一覧 6）。
            返信と、返信の無いメッセージは空
          type: array
          maxItems: 3
          items:
            $ref: '#/components/schemas/UserSummary'
        author:
          description: 投稿者。退会した利用者なら null（削除済みの利用者として表示する。機能一覧 1.5）
          oneOf:
            - $ref: '#/components/schemas/UserSummary'
            - type: 'null'
        body:
          description: 本文。削除済みなら null（本文は DB に残るが返さない。機能一覧 4.2）
          type:
            - string
            - 'null'
        createdAt:
          type: string
          format: date-time
        editedAt:
          description: 最後に編集した時刻。編集していなければ null（「編集済み」の表示に使う。機能一覧 4.2）
          type:
            - string
            - 'null'
          format: date-time
        deleted:
          description: 削除済みか（「このメッセージは削除されました」に置き換えて表示する。機能一覧 4.2）
          type: boolean
        mentions:
          description: >-
            本文のメンションの対象（機能一覧 9.1）。投稿・返信のときに、そのチャンネルの参加者で退会していない利用者へ解決できた
            `@ユーザーID` を保存し、本文に最初に現れた順に1人1件で載せる。編集では、本文にユーザーID が残っている対象は解決し直さずに残す
            （チャンネルを抜けた・退会した対象も残り、退会した対象は user が null）。本文から消えた対象は外し、新しく書いた `@` は投稿と同じく解決して足す。
            削除済みのメッセージは空
          type: array
          items:
            $ref: '#/components/schemas/MessageMention'
        reactions:
          description: >-
            リアクション（機能一覧 7）。絵文字ごとに1件で、その絵文字を初めて付けた順。削除済みのメッセージは空
            （本文を返さないのと同じく返さない。付けた記録は DB に残る）
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/Reaction'
        attachments:
          description: >-
            投稿に付けた添付（上げた順）。削除済みのメッセージは空（本文を返さないのと同じく返さない。機能一覧 4.2）
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
    Reaction:
      description: 1つの絵文字のリアクション（機能一覧 7）
      type: object
      additionalProperties: false
      required:
        - emoji
        - count
        - users
      properties:
        emoji:
          description: 絵文字1つ
          type: string
        count:
          description: >-
            付けた人数（カウンタ列。一覧で COUNT を発行しない。要件定義書 4.1）。退会した利用者の分も数える
          type: integer
          minimum: 1
        users:
          description: >-
            付けた人（「誰が付けたか」をホバーで出す。機能一覧 7）。付けた順。退会した利用者は含めない
            （そのため count より少ないことがある）
          type: array
          items:
            $ref: '#/components/schemas/UserSummary'
...
    MessageMention:
      description: メンションの対象（表示時の参照先。機能一覧 9.1 の経路2。対象がいまそのチャンネルの参加者かは問わない）
      type: object
      additionalProperties: false
      required:
        - userId
        - user
      properties:
        userId:
          description: 対象のユーザーID（登録したときの大文字小文字のまま）。本文の `@ユーザーID` とは大文字小文字によらず照合する（機能一覧 1.1）
          type: string
        user:
          description: 対象の利用者。退会した利用者なら null（削除済みの利用者として表示する。機能一覧 1.5）
          oneOf:
            - $ref: '#/components/schemas/UserSummary'
            - type: 'null'
    MessagePage:
      type: object
      additionalProperties: false
      required:
        - messages
        - nextBefore
      properties:
        messages:
          description: 新しい順
          type: array
          items:
            $ref: '#/components/schemas/Message'
        nextBefore:
          description: 続きを取るときに before に渡す id。続きが無ければ null
          type:
            - string
            - 'null'
          format: uuid
...
    StartDmRequest:
      type: object
      additionalProperties: false
      required:
        - userId
      properties:
        userId:
          description: 相手の User.id（ワークスペースの参加者一覧が返す id。ユーザーID ではない）
          type: string
          format: uuid
    UpdateDmReadRequest:
      type: object
      additionalProperties: false
      required:
        - lastReadMessageId
      properties:
        lastReadMessageId:
          description: >-
            読んだ位置（その DM のメッセージの id。F-23。機能一覧 10.1）。
            既読位置は戻さないため、いま持っている位置より古い id を渡しても既読位置は変わらない
          type: string
          format: uuid
    Dm:
      description: 自分が当事者の DM（機能一覧 8・10.1）
      type: object
      additionalProperties: false
      required:
        - id
        - counterpart
        - writable
        - joinedAt
        - unread
        - lastReadMessageId
      properties:
        id:
          type: string
          format: uuid
        counterpart:
          description: 相手。退会した利用者なら null（削除済みの利用者として表示する。機能一覧 1.5）
          oneOf:
            - $ref: '#/components/schemas/UserSummary'
            - type: 'null'
        writable:
          description: >-
            相手がいまこのワークスペースの退会していないメンバーか。false なら投稿できない（409 dm_counterpart_unavailable。機能一覧 8）。
            過去のメッセージは読める
          type: boolean
        joinedAt:
          description: >-
            要求した利用者がこのワークスペースに参加した時刻（F-23。機能一覧 10.1）。既読位置をまだ持たないとき、
            未読はこの時刻より後のメッセージだけを数え、「ここから未読」の区切り線もこの時刻より後の最初のメッセージの上に出す
          type: string
          format: date-time
        unread:
          description: >-
            要求した利用者の未読数（F-23。機能一覧 10.1）。既読位置からの差分で求め、自分の投稿と削除済みは数えない。
            既読位置をまだ持たない利用者は、joinedAt より後だけを数える
          type: integer
          minimum: 0
        lastReadMessageId:
          description: >-
            要求した利用者の既読位置（F-23。機能一覧 10.1）。「ここから未読」の区切り線を、この次のメッセージの上に出す。
            まだ既読位置を持たなければ null。未読数から線の位置を数えてはならない
          type:
            - string
            - 'null'
          format: uuid
    DmMessage:
      description: DM のメッセージ（機能一覧 8・4.2）
      type: object
      additionalProperties: false
      required:
        - id
        - dmId
        - author
        - body
        - createdAt
        - editedAt
        - deleted
      properties:
        id:
          description: UUIDv7。一覧のカーソル（before）にそのまま使う
          type: string
          format: uuid
        dmId:
          type: string
          format: uuid
        author:
          description: 書き手。退会した利用者なら null（削除済みの利用者として表示する。機能一覧 1.5）
          oneOf:
            - $ref: '#/components/schemas/UserSummary'
            - type: 'null'
        body:
          description: 本文。削除済みなら null（本文は DB に残るが返さない。機能一覧 4.2）
          type:
            - string
            - 'null'
        createdAt:
          type: string
          format: date-time
        editedAt:
          description: 最後に編集した時刻。編集していなければ null（機能一覧 4.2）
          type:
            - string
            - 'null'
          format: date-time
        deleted:
          description: 削除済みか（機能一覧 4.2）
          type: boolean
    DmMessagePage:
      type: object
      additionalProperties: false
      required:
        - messages
        - nextBefore
      properties:
        messages:
          description: 新しい順
          type: array
          items:
            $ref: '#/components/schemas/DmMessage'
        nextBefore:
          description: 続きを取るときに before に渡す id。続きが無ければ null
          type:
            - string
            - 'null'
          format: uuid
...